requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
requests>=2.26.0
python-dotenv>=0.19.0
web3>=5.24.0
python-multipart>=0.0.5
pinata-sdk>=0.1.0
pydantic>=1.8.2
typing-extensions>=4.0.0
eth-account>=0.5.9
PyJWT>=2.3.0
cryptography>=3.4.8
python-jose>=3.3.0
passlib>=1.7.4
beacon-chain>=0.1.0
aiohttp>=3.8.1